Phase 6 close-out: address review findings on translator + parser
Translator hardening (hailo_cs_translator.c):
- Fix build break: translate_dynamic call site now passes cfg.
- translate_dynamic refuses to run on any truncated flag
  (ctx->truncated or per-kind *_truncated) instead of silently
  emitting a wrong-order stream when cursor positions drift.
- Unknown action tags in action_types[] now return HAILO_ERR_INVAL
  rather than silently skipping — firmware can't detect "action
  omitted," so loud failure is the only safe choice.
- translate_allow_input_dataflow fails cleanly when sys_index isn't
  present in info->pads[] (previously emitted frame_periph_size=0)
  and guards the (config_vdma_channel + 1) → u8 narrowing.
- Magic proto-tag literals 8/7/5/6/10 replaced with
  ProtoHEFAction_*_tag constants.

Parser hardening (hef_parser.{c,h}):
- Guard 1u << field->tag against UB for any future proto branch
  past 31. Clamp with explicit < 32 check.
- Promote context_index from uint8_t to uint32_t on all five new
  action structs; source value in hef_context_actions is u32 so
  the narrowing silently aliased mappings at >255 contexts (latent
  until MAX_CONTEXTS grows).
- Saturate per-kind *_count at MAX when truncated, matching the
  pad_count convention — callers can index [0..count) safely.

Control-channel MSI race (hailo_control.c):
- Clear control_msi_pending AFTER the doorbell write + mb(), not
  before. A stale MSI arriving in the pre-doorbell window would
  otherwise set pending=1 and cause wait_for_response to read a
  prior RPC's buffer (→ the 0xFFFFFFFF buffer_len mode on pi-5-1).

Nanopb oneof drain clarification:
- Comment at the decode_compute_action_inner_cb tail explains why
  pb_read(stream, NULL, stream->bytes_left) drains the per-branch
  substream, not the parent ProtoHEFAction — addressing a reviewer
  concern that turned out to be a false flag.

Tests (test_hailo.c + test_hef_parser.c):
- Byte-for-byte wire-format tests for all 4 new translator
  functions (disable_lcu, wait_sequencer, trigger_sequencer,
  allow_input_dataflow) — each with sizeof(body) + field offsets
  checked.
- Interleaved-kinds order test: action_types[5] with mixed tags
  verifies the per-kind cursor dispatch preserves HEF sequence.
- Unknown-tag and truncated-context failure tests for the
  translator's new error paths.
- Malformed-body rejection tests for the 4 new parser action
  kinds (truncated varint → pb_decode returns false).
- enable_sequencer cap-overflow test matching the existing
  pads/enable_lcu truncation pattern.
- Existing enable_lcu translator tests updated to populate
  context_actions[0].action_types[] (previously implicit under
  the pre-review translate_dynamic walk).

Built clean on QEMU-ARM64, RASPI5, JETSON_ORIN_NANO, X86_64.
All tests pass under make test AI_SCHED=ON.

+++ kernel/ai_accel/hailo/hef_parser.h
@@ -207,7 +207,7 @@ struct hef_context_actions {
 #define HEF_PARSER_MAX_ENABLE_LCU_ACTIONS  32u
 
 struct hef_enable_lcu_action {
-    uint8_t  context_index;
+    uint32_t context_index;
     uint32_t lcu_index;
     uint32_t cluster_index;
     uint32_t network_index;
@@ -225,7 +225,7 @@ struct hef_enable_lcu_action {
 #define HEF_PARSER_MAX_DISABLE_LCU_ACTIONS  32u
 
 struct hef_disable_lcu_action {
-    uint8_t  context_index;
+    uint32_t context_index;
     uint32_t lcu_index;
     uint32_t cluster_index;
     uint32_t lcu_enable_address;
@@ -244,7 +244,7 @@ struct hef_disable_lcu_action {
 #define HEF_PARSER_MAX_TRIGGER_SEQUENCER_ACTIONS  16u
 
 struct hef_trigger_sequencer_action {
-    uint8_t  context_index;
+    uint32_t context_index;
     uint32_t cluster_index;
     uint32_t initial_l3_cut;
     uint32_t initial_l3_offset;
@@ -265,7 +265,7 @@ struct hef_trigger_sequencer_action {
 #define HEF_PARSER_MAX_WAIT_SEQUENCER_ACTIONS  16u
 
 struct hef_wait_sequencer_action {
-    uint8_t  context_index;
+    uint32_t context_index;
     uint32_t cluster_index;
 };
 
@@ -281,7 +281,7 @@ struct hef_wait_sequencer_action {
 #define HEF_PARSER_MAX_ALLOW_INPUT_DATAFLOW_ACTIONS  8u
 
 struct hef_allow_input_dataflow_action {
-    uint8_t  context_index;
+    uint32_t context_index;
     uint32_t sys_index;
     uint32_t connection_type;
 };
